added tests, completed some TODOs.
This commit refs #18488.

+++ src/main/java/buildingsmart/ifc/IfcDirection.java
@@ -87,7 +87,7 @@ public class IfcDirection extends IfcGeometricRepresentationItem
             directionRatiosList.add(new IfcReal(dirRatio));
         }
         this.directionRatios = directionRatiosList;
-        this.dim = new IfcDimensionCount((byte) directionRatiosList.size());
+        this.dim = new IfcDimensionCount(directionRatiosList.size());
     }
 
     /**
